requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
numpy==1.15.3
openapi-codec==1.3.2
openpyxl==2.5.9
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.7.4
python-decouple==3.1
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
